requirements.txt
python-telegram-bot==20.3
httpx
python-dotenv